requirements.txt
# Core Bot Framework
python-telegram-bot==20.7

# Download Engine - Updated to latest
yt-dlp==2025.9.5

# HTTP and Async Support
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.2

# Image Processing
Pillow==10.1.0

# Configuration Management
python-dotenv==1.0.0

# Audio/Video Processing (for MP3 conversion)
mutagen==1.47.0
pycryptodomex==3.23.0

# Compression and Encoding
brotli==1.1.0

# Async Support
asyncio-mqtt==0.16.1

# Logging and Utilities
colorama==0.4.6

# Optional: FFmpeg Python wrapper (if needed)
# ffmpeg-python==0.2.0

# Development and Testing (optional)
# pytest==7.4.3
# pytest-asyncio==0.21.1
